finish task01

+++ src/main/java/edu/chat/models/Chatroom.java
@@ -9,12 +9,12 @@ import java.util.List;
 @ToString
 @EqualsAndHashCode
 public class Chatroom {
-    @Getter private final Long id;
+    @Getter private final int id;
     @Getter private final String name;
     @Getter private final User owner;
     @Getter private List<Message> chatroomMessages;
 
-    public Chatroom(Long id, String name, User owner) {
+    public Chatroom(int id, String name, User owner) {
         this.id = id;
         this.name = name;
         this.owner = owner;
